requirements.txt
crewai==0.22.5
crewai[tools]
langchain_openai
langchain_community
python-dotenv
requests
PyPDF2
pdfplumber
